requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.51.0
packaging==24.1
pathspec==0.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyYAML==6.0.2
setuptools==75.1.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
